fix(trips): make announcement links clickable
Mutualize URL linkification/opening between chat and announcements so admin announcements now behave like messaging links with consistent styling and tap handling.

Made-with: Cursor
